fix: Added the proper types
